[events] Simplify getEvent in EventModule\BasePresenter
The current logic is too convoluted and explits Presenter::redirect to
dispatch users who did not specify eventId. That is a required
parameter, so it makes sense to always throw BadReqeustException when
it's not filled it.

Doing the redirect (via AbortException) also breaks
BasePresenter::authorized() where only BadRequestException makes sense.